Remove old context
