--- burn-wgpu/src/context/mod.rs
@@ -1,6 +1,0 @@
-pub(super) mod client;
-pub(super) mod server;
-
-mod base;
-
-pub use base::*;
