+++ burn-wgpu/src/kernel/pool/avg_pool2d.rs
@@ -1,4 +1,5 @@
 use crate::{
+    compute::{Kernel, StaticKernel},
     element::WgpuElement,
     kernel::{
         self, elemwise_workgroup,
@@ -39,22 +40,21 @@ pub(crate) fn avg_pool2d<E: WgpuElement>(
 ) -> WgpuTensor<E, 4> {
     const WORKGROUP: usize = 32;
 
-    let (info_buffer, output) =
+    let (info_handle, output) =
         build_output_and_info_pool2d(&x, kernel_size, stride, padding, [1, 1]);
-    let kernel = match count_include_pad {
-        true => x
-            .context
-            .compile_static::<KernelSettings<AvgPool2d<true>, E, i32, WORKGROUP, WORKGROUP, 1>>(),
-        false => x
-            .context
-            .compile_static::<KernelSettings<AvgPool2d<false>, E, i32, WORKGROUP, WORKGROUP, 1>>(),
+
+    let workgroup = elemwise_workgroup(output.shape.num_elements(), WORKGROUP);
+    let kernel: Box<dyn Kernel> = match count_include_pad {
+        true => Box::new(StaticKernel::<
+            KernelSettings<AvgPool2d<true>, E, i32, WORKGROUP, WORKGROUP, 1>,
+        >::new(workgroup)),
+        false => Box::new(StaticKernel::<
+            KernelSettings<AvgPool2d<false>, E, i32, WORKGROUP, WORKGROUP, 1>,
+        >::new(workgroup)),
     };
 
-    x.context.execute(
-        elemwise_workgroup(output.shape.num_elements(), WORKGROUP),
-        kernel,
-        &[&x.buffer, &output.buffer, &info_buffer],
-    );
+    x.client
+        .execute(kernel, &[&x.handle, &output.handle, &info_handle]);
 
     output
 }
@@ -72,37 +72,22 @@ pub(crate) fn avg_pool2d_backward<E: WgpuElement>(
     let grad = kernel::into_contiguous(grad);
 
     let num_elems = x.shape.num_elements();
-    let buffer = x
-        .context
-        .create_buffer(num_elems * core::mem::size_of::<E>());
-    let output = WgpuTensor::new(x.context.clone(), x.shape.clone(), buffer);
-    let info_buffer = build_pool2d_info(&x, &grad, kernel_size, stride, padding, [1, 1]);
+    let buffer = x.client.empty(num_elems * core::mem::size_of::<E>());
+    let output = WgpuTensor::new(x.client.clone(), x.device, x.shape.clone(), buffer);
+    let info_handle = build_pool2d_info(&x, &grad, kernel_size, stride, padding, [1, 1]);
+    let workgroup = elemwise_workgroup(output.shape.num_elements(), WORKGROUP);
 
-    let kernel =
-        match count_include_pad {
-            true => x.context.compile_static::<KernelSettings<
-                AvgPool2dBackward<true>,
-                E,
-                i32,
-                WORKGROUP,
-                WORKGROUP,
-                1,
-            >>(),
-            false => x.context.compile_static::<KernelSettings<
-                AvgPool2dBackward<false>,
-                E,
-                i32,
-                WORKGROUP,
-                WORKGROUP,
-                1,
-            >>(),
-        };
+    let kernel: Box<dyn Kernel> = match count_include_pad {
+        true => Box::new(StaticKernel::<
+            KernelSettings<AvgPool2dBackward<true>, E, i32, WORKGROUP, WORKGROUP, 1>,
+        >::new(workgroup)),
+        false => Box::new(StaticKernel::<
+            KernelSettings<AvgPool2dBackward<false>, E, i32, WORKGROUP, WORKGROUP, 1>,
+        >::new(workgroup)),
+    };
 
-    x.context.execute(
-        elemwise_workgroup(output.shape.num_elements(), WORKGROUP),
-        kernel,
-        &[&grad.buffer, &output.buffer, &info_buffer],
-    );
+    x.client
+        .execute(kernel, &[&grad.handle, &output.handle, &info_handle]);
 
     output
 }
