+++ burn-wgpu/src/kernel/prng/base.rs
@@ -5,7 +5,10 @@ use burn_tensor::Shape;
 use rand::Rng;
 use wgpu::Buffer;
 
-use crate::{context::Context, element::WgpuElement, kernel_wgsl, tensor::WgpuTensor, SEED};
+use crate::{
+    compute::WgpuComputeClient, element::WgpuElement, kernel_wgsl, tensor::WgpuTensor, WgpuDevice,
+    SEED,
+};
 
 kernel_wgsl!(Prng, "../../template/prng/prng.wgsl");
 
@@ -24,21 +27,28 @@ pub(crate) fn get_seeds() -> Vec<u32> {
 }
 
 pub(crate) fn make_output_tensor<E: WgpuElement, const D: usize>(
-    context: Arc<Context>,
+    client: WgpuComputeClient,
+    device: WgpuDevice,
     shape: Shape<D>,
 ) -> WgpuTensor<E, D> {
-    let buffer = context.create_buffer(shape.num_elements() * core::mem::size_of::<E>());
-    WgpuTensor::new(context, shape, buffer)
+    let buffer = client.empty(shape.num_elements() * core::mem::size_of::<E>());
+    WgpuTensor::new(client, device, shape, buffer)
 }
 
-pub(crate) fn make_info_buffer(context: Arc<Context>, n_values_per_thread: usize) -> Arc<Buffer> {
+pub(crate) fn make_info_buffer(
+    client: WgpuComputeClient,
+    n_values_per_thread: usize,
+) -> Arc<Buffer> {
     let mut info = get_seeds();
     info.insert(0, n_values_per_thread as u32);
-    context.create_buffer_with_data(bytemuck::cast_slice(&info))
+    client.create(bytemuck::cast_slice(&info))
 }
 
-pub(crate) fn make_args_buffer<E: WgpuElement>(context: Arc<Context>, args: &[E]) -> Arc<Buffer> {
-    context.create_buffer_with_data(E::as_bytes(args))
+pub(crate) fn make_args_buffer<E: WgpuElement>(
+    client: WgpuComputeClient,
+    args: &[E],
+) -> Arc<Buffer> {
+    client.create(E::as_bytes(args))
 }
 
 #[cfg(test)]
