+++ burn-wgpu/src/kernel/prng/normal.rs
@@ -1,12 +1,12 @@
 use burn_tensor::Shape;
 
 use crate::{
+    compute::{compute_client, StaticKernel},
     element::WgpuElement,
     kernel::{
         prng::base::{make_args_buffer, make_info_buffer, make_output_tensor},
         prng_workgroup, KernelSettings, SourceTemplate, StaticKernelSource,
     },
-    pool::get_context,
     tensor::WgpuTensor,
     GraphicsApi, WgpuDevice,
 };
@@ -39,15 +39,17 @@ pub fn random_normal<G: GraphicsApi, E: WgpuElement, const D: usize>(
     const WORKGROUP: usize = 32;
     const N_VALUES_PER_THREAD: usize = 128; // must be even
 
-    let context = get_context::<G>(device);
-    let output = make_output_tensor(context.clone(), shape.clone());
-    let info_buffer = make_info_buffer(context.clone(), N_VALUES_PER_THREAD);
-    let args_buffer = make_args_buffer(context.clone(), &[mean, std]);
+    let client = compute_client::<G>(device);
+    let output = make_output_tensor(client.clone(), device.clone(), shape.clone());
+    let info_handle = make_info_buffer(client.clone(), N_VALUES_PER_THREAD);
+    let args_handle = make_args_buffer(client.clone(), &[mean, std]);
+    let workgroup = prng_workgroup(shape.num_elements(), WORKGROUP, N_VALUES_PER_THREAD);
+    let kernel =
+        StaticKernel::<KernelSettings<NormalPrng, E, i32, WORKGROUP, WORKGROUP, 1>>::new(workgroup);
 
-    context.execute(
-        prng_workgroup(shape.num_elements(), WORKGROUP, N_VALUES_PER_THREAD),
-        context.compile_static::<KernelSettings<NormalPrng, E, i32, WORKGROUP, WORKGROUP, 1>>(),
-        &[&output.buffer, &info_buffer, &args_buffer],
+    client.execute(
+        Box::new(kernel),
+        &[&output.handle, &info_handle, &args_handle],
     );
 
     output
