+++ burn-wgpu/src/lib.rs
@@ -12,15 +12,10 @@ mod ops;
 pub mod benchmark;
 /// Compute related module.
 pub mod compute;
-/// Context module.
-pub mod context;
 /// Kernel module
 pub mod kernel;
 /// Tensor module.
 pub mod tensor;
-
-pub(crate) mod pool;
-pub(crate) mod tune;
 
 mod element;
 pub use element::{FloatElement, IntElement};
