--- burn-wgpu/src/tune/mod.rs
@@ -1,3 +1,0 @@
-mod base;
-
-pub use base::*;
